refactor : BestSeller component
기존 main 내 BestSeller component를 공용으로 사용하기 위한 refactoring 작업

Related to : #58
