+++ src/app/page.tsx
@@ -2,11 +2,11 @@ import RecentlyViewedBooks from '@/components/layout/RecentlyViewedBooks';
 import styles from '@/styles/main/main.module.css';
 import NewBook from '../components/main/newBook/NewBook';
 import ThemeRecommendation from '@/components/main/themeRecommendation/ThemeRecommendation';
-import BestSeller from '@/components/main/bestSeller/BestSeller';
 import UsedBook from '@/components/main/usedBook/UsedBook';
 import ImageSlider from '@/components/main/imageSlider/ImageSlider';
 import SignupModal from '@/components/auth/signup/SignupModal';
 import LoginModal from '@/components/auth/login/LoginModal';
+import BestSeller from '@/components/common/BestSeller';
 
 export default function Home() {
 	return (
